libnm-core: add support to first level only json get/set align functions
In team properties there are no simple values to be assigned to first
level keys: generally all the keys are contained in the "runner" key.
In team-port json config instead almost all the properties are single,
first level keys. Add support to this missing case in nm-utils.c json
align functions.
In the meanwhile, rework a bit _nm_utils_team_config_set to make search
and del/update of the keys more symmetrical (and more clear).